move some hooks to core
